backend/requirements.txt
Django==4.2.7
djangorestframework==3.15.1
mysqlclient==2.2.4
python-dotenv==1.0.1
pytest==8.2.0
pytest-django==4.8.0
drf-spectacular==0.27.2
django-filter==24.2
